Separated out Bound from Signature.
