+++ everest/cascade/__init__.py
@@ -11,6 +11,7 @@ from .. import (
 from .hierarchy import *
 from .cascade import *
 from .signature import *
+from .bound import *
 from .params import *
